backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic[email]
redis==5.2.1
httpx==0.28.1
python-slugify==8.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
stripe==8.4.0
# Test dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
httpx[http2]
